+++ src/layertree/component.js
@@ -15,7 +15,7 @@ Controller.$inject = [
 ];
 // The MIT License (MIT)
 //
-// Copyright (c) 2016-2024 Camptocamp SA
+// Copyright (c) 2016-2025 Camptocamp SA
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
@@ -55,6 +55,8 @@ import ngeoLayertreeController, {LayertreeVisitorDecision} from 'ngeo/layertree/
 import ngeoMapLayerHelper from 'ngeo/map/LayerHelper';
 import ngeoMiscSyncArrays from 'ngeo/misc/syncArrays';
 import ngeoMiscWMSTime from 'ngeo/misc/WMSTime';
+import 'gmf/time-input/timeslider';
+import 'gmf/time-input/datepicker';
 import olLayerTile from 'ol/layer/WebGLTile';
 import olLayerLayer from 'ol/layer/Layer';
 import {isEmpty} from 'ol/obj';
@@ -346,6 +348,12 @@ export function Controller(
    */
   this.$timeout_ = $timeout;
 
+  /**
+   * updateTimeData with bound context.
+   * @type {Controller.prototype.updateTimeData}
+   */
+  this.updateTimeDataBind = this.updateTimeData.bind(this);
+
   // enter digest cycle on node collapse
   $element.on('shown.bs.collapse', () => {
     this.scope_.$apply();
@@ -533,11 +541,11 @@ Controller.prototype.getNodeState = function (treeCtrl) {
  *
  * LayertreeController.prototype.updateTimeData - description
  *
- * @param {import('ngeo/layertree/Controller').LayertreeController} layertreeCtrl ngeo layertree controller
  * @param {import('ngeo/datasource/OGC').TimeRange} time The start
+ * @param {import('ngeo/layertree/Controller').LayertreeController} layertreeCtrl ngeo layertree controller
  * and optionally the end datetime (for time range selection) selected by user
  */
-Controller.prototype.updateTimeData = function (layertreeCtrl, time) {
+Controller.prototype.updateTimeData = function (time, layertreeCtrl) {
   this.updateWMSTimeLayerState(layertreeCtrl, time);
   this.gmfPermalink_.refreshLayerTime(layertreeCtrl, time);
 };