+++ CHANGES.md
@@ -15,6 +15,8 @@ can be compiled with any modern bundler.
 - Let the users choose the version of peerDependencies
 - Remove not used offline and ol-cesium related code.
 - Move jquery and jquery-ui import out of components.
+- Remove ngeo-date-picker, replace it by gmf-datepicker
+- Remove gmf-date-slider, replace it by gmf-timeslider
 
 # 2.8